requirements.txt
argh==0.26.2
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
pathtools==0.1.2
Pillow==6.2.1
PyYAML==5.2
watchdog==0.9.0
Werkzeug==0.16.0
